pyproject.toml
[project]
name = "coresupportapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.1",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-csv>=3.0.2",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.7",
    "paramiko>=4.0.0",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy==2.0.41",
    "uvicorn>=0.37.0",
]
